+++ components/st-vertical-tabs/index.ts
@@ -1,0 +1,1 @@
+export * from './st-vertical-tabs.module';