+++ components/st-vertical-tabs/st-vertical-tabs.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { StVerticalTabsComponent } from './st-vertical-tabs.component';
+
+@NgModule({
+   imports: [CommonModule],
+   declarations: [StVerticalTabsComponent],
+   exports: [StVerticalTabsComponent]
+})
+export class StVerticalTabsModule { }