+++ web/app/+examples/vertical-tabs/index.ts
@@ -1,0 +1,1 @@
+export * from './vertical-tabs.component';