+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Add optional Pre-title to Page Title Component
 * Update color definitions
 * Add util service for manage translations of complex objects
+* Update fonts in sass for use body istead open-sans and heading instead brandon-grotesque
+* Fix some problems with fonts
 * BREAKING CHANGE. Remove ng2-translate dependency
 * BREAKING CHANGE. Now egeo is imported like this: "import { EgeoModule } from 'egeo';"
 * BREAKING CHANGE. Output Search Component change of (onSearch) to (search);